meson: Do not link libGL when compiling for Windows
When building with MinGW meson can find 'GL/gl.h' but there is not libGL
to link to, so compilation eventually fails with the following message:

  /usr/bin/x86_64-w64-mingw32-ld: cannot find -lGL
  collect2: error: ld returned 1 exit status

To avoid that and have cairo build fine with MinGW, do not enable the gl
backed only based on the presence of the header file when compiling for
Windows.

For proper OpenGL detection on Windows meson should probably look for
the opengl32 library, but this change does not got that far.